+++ packages/react/src/components/Toast/index.tsx
@@ -1,0 +1,34 @@
+import { X } from "phosphor-react"
+import { CloseButton, StyledToast, ToastDescription, ToastTitle, ToastViewport } from "./styles"
+import * as ToastPrimitive from '@radix-ui/react-toast'
+
+interface ToastProps {
+  title: string
+  description?: string
+  type?: 'success' | 'error'
+  onClose?: () => void
+}
+
+export function Toast({
+  title,
+  description,
+  type = 'success',
+  onClose,
+}: ToastProps) {
+  return (
+    <ToastPrimitive.Provider>
+      <StyledToast type={type} onOpenChange={(open) => !open && onClose?.()}>
+        <div>
+          <ToastTitle>{title}</ToastTitle>
+          {description && <ToastDescription>{description}</ToastDescription>}
+        </div>
+        <CloseButton asChild>
+          <button aria-label="Close">
+            <X size={20} />
+          </button>
+        </CloseButton>
+      </StyledToast>
+      <ToastViewport />
+    </ToastPrimitive.Provider>
+  )
+}