+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @alfa-ui/tokens
 
+## 3.0.0
+
+### Major Changes
+
+- Introduce a new component - Toast and uploado for npm
+
 ## 2.0.0
 
 ### Major Changes